requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

numpy
pillow
pandas
scikit-learn
matplotlib

streamlit
opencv-python
